+++ sim/rotation/requirements.txt
@@ -1,0 +1,58 @@
+absl-py==2.3.1
+cloudpickle==3.1.1
+contourpy==1.3.2
+cycler==0.12.1
+etils==1.13.0
+Farama-Notifications==0.0.4
+filelock==3.18.0
+fonttools==4.59.0
+fsspec==2025.7.0
+glfw==2.9.0
+grpcio==1.73.1
+gymnasium==1.2.0
+imageio==2.37.0
+imageio-ffmpeg==0.6.0
+importlib_resources==6.5.2
+Jinja2==3.1.6
+kiwisolver==1.4.8
+Markdown==3.8.2
+MarkupSafe==3.0.2
+matplotlib==3.10.3
+mpmath==1.3.0
+mujoco==3.3.4
+networkx==3.4.2
+numpy==2.2.6
+nvidia-cublas-cu12==12.6.4.1
+nvidia-cuda-cupti-cu12==12.6.80
+nvidia-cuda-nvrtc-cu12==12.6.77
+nvidia-cuda-runtime-cu12==12.6.77
+nvidia-cudnn-cu12==9.5.1.17
+nvidia-cufft-cu12==11.3.0.4
+nvidia-cufile-cu12==1.11.1.6
+nvidia-curand-cu12==10.3.7.77
+nvidia-cusolver-cu12==11.7.1.2
+nvidia-cusparse-cu12==12.5.4.2
+nvidia-cusparselt-cu12==0.6.3
+nvidia-nccl-cu12==2.26.2
+nvidia-nvjitlink-cu12==12.6.85
+nvidia-nvtx-cu12==12.6.77
+packaging==25.0
+pandas==2.3.1
+pillow==11.3.0
+protobuf==6.31.1
+PyOpenGL==3.1.9
+pyparsing==3.2.3
+python-dateutil==2.9.0.post0
+pytz==2025.2
+PyYAML==6.0.2
+scipy==1.15.3
+six==1.17.0
+sympy==1.14.0
+tensorboard==2.20.0
+tensorboard-data-server==0.7.2
+torch==2.7.1
+triton==3.3.1
+typing_extensions==4.14.1
+tzdata==2025.2
+Werkzeug==3.1.3
+zipp==3.23.0
